requirements.txt
setuptools~=70.1.0
python-dateutil~=2.9.0
pandas~=1.5.3
regex~=2024.5.15
numpy~=1.26.4
psycopg2~=2.9.9
dataclasses~=0.8
marshmallow~=3.21.3
yaml~=0.2.5
pyyaml~=6.0.1
sympy~=1.12.1